+++ src/components/common/gettingStartedCard/GettingStarted.tsx
@@ -1,15 +1,9 @@
 import React from 'react'
 import GettingToast from '../../../assets/img/lifebuoy.png'
 import { updateLoginCount } from '../../../services/service'
+import { GettingStartedType } from '../guidePage/onboarding.type'
 import { setActionWithExpiry } from '../helpers/Helpers'
 import './gettingStarted.scss'
-
-export interface GettingStartedType {
-    className: string
-    showHelpCard: boolean
-    hideGettingStartedCard: () => void
-    loginCount: number
-}
 
 function GettingStartedCard({ className, hideGettingStartedCard }: GettingStartedType) {
     const onClickedOkay = () => {
@@ -18,7 +12,7 @@ function GettingStartedCard({ className, hideGettingStartedCard }: GettingStarte
     }
 
     const onClickedDontShowAgain = () => {
-        let updatedPayload = {
+        const updatedPayload = {
             key: 'login-count',
             value: '5',
         }
@@ -29,7 +23,7 @@ function GettingStartedCard({ className, hideGettingStartedCard }: GettingStarte
     return (
         <div className="getting_tippy__position">
             <div className="arrow-up" />
-            <div className={`${className} getting-started-card cn-0 p-20 br-8 fs-13 `}>
+            <div className={`getting-started-card cn-0 p-20 br-8 fs-13 ${className}`}>
                 <img className="mb-12 icon-dim-32" src={GettingToast} alt="getting started icon" />
                 <div className="flex column left fw-6">Getting started</div>
                 <div>You can always access the Getting Started guide from here.</div>
