+++ src/components/common/guidePage/onboarding.type.ts
@@ -1,0 +1,16 @@
+export interface GettingStartedType {
+  className: string
+  showHelpCard: boolean
+  hideGettingStartedCard: () => void
+  loginCount: number
+}
+
+export interface GuidedPageType {
+  openDevtronAppCreateModel: (event) => void
+}
+
+export interface OnboardingGuideProps {
+  loginCount: number
+  isSuperAdmin: boolean
+  serverMode: string
+}
