+++ src/components/common/navigation/navigation.type.ts
@@ -1,0 +1,4 @@
+export interface AppRouterType {
+  isSuperAdmin?: boolean
+  appListCount: number
+}
